backend/requirements.txt
Flask
tensorflow
opencv-python-headless
Pillow
numpy
